requirements.txt
streamlit==1.31.0
pandas==2.2.0
groq==0.5.0
python-dotenv==1.0.0
plotly==5.18.0
yfinance==0.2.37
sqlite3